Document-intelligence-refinery/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "Production-grade, multi-stage agentic pipeline for document extraction with provenance tracking"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "FDE Program", email = "fde@example.com"}
]
keywords = ["document-intelligence", "pdf-extraction", "rag", "provenance", "agentic-pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Markup",
]

dependencies = [
    # Core dependencies
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typing-extensions>=4.8.0",
    
    # Document processing
    "pdfplumber>=0.10.0",
    "pymupdf>=1.23.0",
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    
    # ML/AI dependencies
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.20",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.5.0",
    
    # Vector stores
    "chromadb>=0.4.0",
    "sentence-transformers>=2.3.0",
    
    # Data processing
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    
    # Web framework
    "streamlit>=1.28.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.7.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "ruff>=0.1.6",
    "pre-commit>=3.5.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
]
